feat: hook useOutsideClick
